refactor: 컴퓨터용 함수 분리
